+++ app/smartbrain_3000/web/_app/immutable/chunks/C-w34g7-.js
@@ -1,0 +1,1 @@
+import{ct as e}from"./Dpf1ywhm.js";e();